+++ astropy/wcs/utils.py
@@ -2,17 +2,20 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import numpy as np
+
 from .. import units as u
 from ..extern.six.moves import range
 
-__doctest_skip__ = ['wcs_to_celestial_frame']
+from .wcs import WCS, WCSSUB_CELESTIAL
 
-__all__ = ['add_stokes_axis_to_wcs',
-           'custom_frame_mappings',
+__doctest_skip__ = ['wcs_to_celestial_frame', 'celestial_frame_to_wcs']
+
+__all__ = ['add_stokes_axis_to_wcs', 'celestial_frame_to_wcs',
            'wcs_to_celestial_frame', 'proj_plane_pixel_scales',
            'proj_plane_pixel_area', 'is_proj_plane_distorted',
            'non_celestial_pixel_scales', 'skycoord_to_pixel',
-           'pixel_to_skycoord']
+           'pixel_to_skycoord', 'custom_wcs_to_frame_mappings',
+           'custom_frame_to_wcs_mappings']
 
 
 def add_stokes_axis_to_wcs(wcs, add_before_ind):
@@ -43,9 +46,11 @@ def add_stokes_axis_to_wcs(wcs, add_before_ind):
 
 def _wcs_to_celestial_frame_builtin(wcs):
 
+    # Import astropy.coordinates here to avoid circular imports
     from ..coordinates import FK4, FK4NoETerms, FK5, ICRS, Galactic
+
+    # Import astropy.time here otherwise setup.py fails before extensions are compiled
     from ..time import Time
-    from . import WCSSUB_CELESTIAL
 
     # Keep only the celestial part of the axes
     wcs = wcs.sub([WCSSUB_CELESTIAL])
@@ -95,10 +100,51 @@ def _wcs_to_celestial_frame_builtin(wcs):
     return frame
 
 
+def _celestial_frame_to_wcs_builtin(frame, projection='TAN'):
+
+    # Import astropy.coordinates here to avoid circular imports
+    from ..coordinates import BaseRADecFrame, FK4, FK4NoETerms, FK5, ICRS, Galactic
+
+    # Create a 2-dimensional WCS
+    wcs = WCS(naxis=2)
+
+    if isinstance(frame, BaseRADecFrame):
+
+        xcoord = 'RA--'
+        ycoord = 'DEC-'
+        if isinstance(frame, ICRS):
+            wcs.wcs.radesys = 'ICRS'
+        elif isinstance(frame, FK4NoETerms):
+            wcs.wcs.radesys = 'FK4-NO-E'
+            wcs.wcs.equinox = frame.equinox.byear
+        elif isinstance(frame, FK4):
+            wcs.wcs.radesys = 'FK4'
+            wcs.wcs.equinox = frame.equinox.byear
+        elif isinstance(frame, FK5):
+            wcs.wcs.radesys = 'FK5'
+            wcs.wcs.equinox = frame.equinox.jyear
+        else:
+            return None
+    elif isinstance(frame, Galactic):
+        xcoord = 'GLON'
+        ycoord = 'GLAT'
+    else:
+        return None
+
+    wcs.wcs.ctype[0] = xcoord + '-' + projection
+    wcs.wcs.ctype[1] = ycoord + '-' + projection
+
+    # Make sure that e.g. LONPOLE and other parameters are set
+    wcs.wcs.set()
+
+    return wcs
+
+
 WCS_FRAME_MAPPINGS = [[_wcs_to_celestial_frame_builtin]]
+FRAME_WCS_MAPPINGS = [[_celestial_frame_to_wcs_builtin]]
 
 
-class custom_frame_mappings(object):
+class custom_wcs_to_frame_mappings(object):
     def __init__(self, mappings=[]):
         if hasattr(mappings, '__call__'):
             mappings = [mappings]
@@ -109,6 +155,23 @@ class custom_frame_mappings(object):
 
     def __exit__(self, type, value, tb):
         WCS_FRAME_MAPPINGS.pop()
+
+
+# Backward-compatibility
+custom_frame_mappings = custom_wcs_to_frame_mappings
+
+
+class custom_frame_to_wcs_mappings(object):
+    def __init__(self, mappings=[]):
+        if hasattr(mappings, '__call__'):
+            mappings = [mappings]
+        FRAME_WCS_MAPPINGS.append(mappings)
+
+    def __enter__(self):
+        pass
+
+    def __exit__(self, type, value, tb):
+        FRAME_WCS_MAPPINGS.pop()
 
 
 def wcs_to_celestial_frame(wcs):
@@ -135,8 +198,8 @@ def wcs_to_celestial_frame(wcs):
     instance and should return either an instance of a frame, or `None` if no
     matching frame was found. You can register this function temporarily with::
 
-        >>> from astropy.wcs.utils import wcs_to_celestial_frame, custom_frame_mappings
-        >>> with custom_frame_mappings(my_function):
+        >>> from astropy.wcs.utils import wcs_to_celestial_frame, custom_wcs_to_frame_mappings
+        >>> with custom_wcs_to_frame_mappings(my_function):
         ...     wcs_to_celestial_frame(...)
 
     """
@@ -147,6 +210,77 @@ def wcs_to_celestial_frame(wcs):
                 return frame
     raise ValueError("Could not determine celestial frame corresponding to "
                      "the specified WCS object")
+
+
+def celestial_frame_to_wcs(frame, projection='TAN'):
+    """
+    For a given coordinate frame, return the corresponding WCS object.
+
+    Note that the returned WCS object has only the elements corresponding to
+    coordinate frames set (e.g. ctype, equinox, radesys).
+
+    Parameters
+    ----------
+    frame : :class:`~astropy.coordinates.baseframe.BaseCoordinateFrame` subclass instance
+        An instance of a :class:`~astropy.coordinates.baseframe.BaseCoordinateFrame`
+        subclass instance for which to find the WCS
+    projection : str
+        Projection code to use in ctype, if applicable
+
+    Returns
+    -------
+    wcs : :class:`~astropy.wcs.WCS` instance
+        The corresponding WCS object
+
+    Examples
+    --------
+
+    ::
+
+        >>> from astropy.wcs.utils import celestial_frame_to_wcs
+        >>> from astropy.coordinates import FK5
+        >>> frame = FK5(equinox='J2010')
+        >>> wcs = celestial_frame_to_wcs(frame)
+        >>> wcs.to_header()
+        WCSAXES =                    2 / Number of coordinate axes
+        CRPIX1  =                  0.0 / Pixel coordinate of reference point
+        CRPIX2  =                  0.0 / Pixel coordinate of reference point
+        CDELT1  =                  1.0 / [deg] Coordinate increment at reference point
+        CDELT2  =                  1.0 / [deg] Coordinate increment at reference point
+        CUNIT1  = 'deg'                / Units of coordinate increment and value
+        CUNIT2  = 'deg'                / Units of coordinate increment and value
+        CTYPE1  = 'RA---TAN'           / Right ascension, gnomonic projection
+        CTYPE2  = 'DEC--TAN'           / Declination, gnomonic projection
+        CRVAL1  =                  0.0 / [deg] Coordinate value at reference point
+        CRVAL2  =                  0.0 / [deg] Coordinate value at reference point
+        LONPOLE =                180.0 / [deg] Native longitude of celestial pole
+        LATPOLE =                  0.0 / [deg] Native latitude of celestial pole
+        RADESYS = 'FK5'                / Equatorial coordinate system
+        EQUINOX =               2010.0 / [yr] Equinox of equatorial coordinates
+
+
+    Notes
+    -----
+
+    To extend this function to frames not defined in astropy.coordinates, you
+    can write your own function which should take a
+    :class:`~astropy.coordinates.baseframe.BaseCoordinateFrame` subclass
+    instance and a projection (given as a string) and should return either a WCS
+    instance, or `None` if the WCS could not be determined. You can register
+    this function temporarily with::
+
+        >>> from astropy.wcs.utils import celestial_frame_to_wcs, custom_frame_to_wcs_mappings
+        >>> with custom_frame_to_wcs_mappings(my_function):
+        ...     celestial_frame_to_wcs(...)
+
+    """
+    for mapping_set in FRAME_WCS_MAPPINGS:
+        for func in mapping_set:
+            wcs = func(frame, projection=projection)
+            if wcs is not None:
+                return wcs
+    raise ValueError("Could not determine WCS corresponding to the specified "
+                     "coordinate frame.")
 
 
 def proj_plane_pixel_scales(wcs):
@@ -385,9 +519,6 @@ def skycoord_to_pixel(coords, wcs, origin=0, mode='all'):
     astropy.coordinates.SkyCoord.from_pixel
     """
 
-    from .. import units as u
-    from . import WCSSUB_CELESTIAL
-
     if _has_distortion(wcs) and wcs.naxis != 2:
         raise ValueError("Can only handle WCS with distortions for 2-dimensional WCS")
 
@@ -461,8 +592,7 @@ def pixel_to_skycoord(xp, yp, wcs, origin=0, mode='all', cls=None):
     astropy.coordinates.SkyCoord.from_pixel
     """
 
-    from .. import units as u
-    from . import WCSSUB_CELESTIAL
+    # Import astropy.coordinates here to avoid circular imports
     from ..coordinates import SkyCoord, UnitSphericalRepresentation
 
     # we have to do this instead of actually setting the default to SkyCoord